# src/main/resources/contrato/openapi-user-service.yml
openapi: 3.0.3
info:
  title: API de Gestión de Usuarios
  description: >
    Microservicio **User-Service** responsable de la creación, consulta y listado de usuarios.
  version: "1.0.0"
  contact:
    name: Luiggy Yataco
    email: luiggy.yataco@gmail.com
servers:
  - url: http://localhost:8081
    description: Entorno local (perfil dev)
tags:
  - name: Usuarios
    description: Operaciones relacionadas con la gestión de usuarios

paths:
  /bs-usuarios/gestion-usuarios/v1/crear:
    post:
      summary: Crear un nuevo usuario
      tags: [Usuarios]
      operationId: crearUsuario
      requestBody:
        required: true
        description: Datos necesarios para crear un nuevo usuario.
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
              properties:
                name:
                  type: string
                  description: Nombre completo del usuario. Campo obligatorio.
                  example: "Andrea López"
                email:
                  type: string
                  format: email
                  description: Correo electrónico del usuario en formato válido. Campo obligatorio.
                  example: "andrea.lopez@gmail.com"
            examples:
              ejemploExitoso:
                summary: Ejemplo de creación exitosa
                value:
                  name: "Andrea López"
                  email: "andrea.lopez@gmail.com"
              ejemploFormatoInvalido:
                summary: Ejemplo de formato de correo inválido
                value:
                  name: "Carlos Díaz"
                  email: "correo_invalido"

      responses:
        '200':
          description: Usuario creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrearUsuarioResponse'
              examples:
                exito:
                  summary: Usuario creado correctamente
                  value:
                    id: 1
                    name: "Andrea López"
                    email: "andrea.lopez@gmail.com"
        '400':
          description: Error en los datos enviados (Bad Request)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                formatoInvalido:
                  summary: Email con formato incorrecto
                  value:
                    status: 400
                    title: "BAD_REQUEST"
                    detail: "Formato Invalido"
                    instance: "/bs-usuarios/gestion-usuarios/v1/crear"
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                interno:
                  summary: Error inesperado
                  value:
                    status: 500
                    title: "INTERNAL_ERROR"
                    detail: "Error al guardar el usuario en base de datos"
                    instance: "/bs-usuarios/gestion-usuarios/v1/crear"

  /bs-usuarios/gestion-usuarios/v1/listar:
    get:
      summary: Listar todos los usuarios registrados
      tags: [Usuarios]
      operationId: listarUsuarios
      responses:
        '200':
          description: Listado exitoso de usuarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CrearUsuarioResponse'
              examples:
                listado:
                  summary: Ejemplo de listado
                  value:
                    - id: 1
                      name: "Andrea López"
                      email: "andrea.lopez@gmail.com"
                    - id: 2
                      name: "Carlos Díaz"
                      email: "carlos.diaz@outlook.com"
        '404':
          description: No existen usuarios registrados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                vacio:
                  summary: No se encontraron usuarios
                  value:
                    status: 404
                    title: "NOT_FOUND"
                    detail: "No existen usuarios registrados"
                    instance: "/bs-usuarios/gestion-usuarios/v1/listar"
        '500':
          description: Error interno al obtener los usuarios
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                interno:
                  summary: Fallo inesperado
                  value:
                    status: 500
                    title: "INTERNAL_ERROR"
                    detail: "Error inesperado en la base de datos"
                    instance: "/bs-usuarios/gestion-usuarios/v1/listar"

  /bs-usuarios/gestion-usuarios/v1/{id}:
    get:
      summary: Obtener un usuario por su identificador
      tags: [Usuarios]
      operationId: obtenerPorId
      parameters:
        - name: id
          in: path
          required: true
          description: ID único del usuario
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Usuario encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrearUsuarioResponse'
              examples:
                encontrado:
                  summary: Usuario encontrado
                  value:
                    id: 1
                    name: "Andrea López"
                    email: "andrea.lopez@gmail.com"
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                noEncontrado:
                  summary: El usuario no existe
                  value:
                    status: 404
                    title: "NOT_FOUND"
                    detail: "Usuario no encontrado"
                    instance: "/bs-usuarios/gestion-usuarios/v1/99"
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                interno:
                  summary: Error inesperado
                  value:
                    status: 500
                    title: "INTERNAL_ERROR"
                    detail: "Error al consultar usuario"
                    instance: "/bs-usuarios/gestion-usuarios/v1/99"

components:
  schemas:
    CrearUsuarioRequest:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
          example: "Andrea López"
        email:
          type: string
          format: email
          example: "andrea.lopez@gmail.com"

    CrearUsuarioResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Andrea López"
        email:
          type: string
          example: "andrea.lopez@gmail.com"

    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          example: 400
        title:
          type: string
          example: "BAD_REQUEST"
        detail:
          type: string
          example: "Formato Invalido"
        instance:
          type: string
          example: "/bs-usuarios/gestion-usuarios/v1/crear"